Update render.sh so it can take an arg: man to render.

+++ render.sh
@@ -1,9 +1,16 @@
 #!/usr/bin/env bash
 
-rm renders/*
+function renderFile {
+  echo $1
+  ./manner.php $1 > /tmp/manner-render-sh && $HOME/bin/tidy-html5/build/cmake/tidy -indent -omit -utf8 --tidy-mark no --wrap 0 -q /tmp/manner-render-sh > renders/$(basename $1).html
+}
 
-for f in ../manlib/man/man1/*
-do
-  echo ${f}
-  ./manner.php ${f} > /tmp/manner-render-sh && $HOME/bin/tidy-html5/build/cmake/tidy -indent -omit -utf8 --tidy-mark no --wrap 0 -q /tmp/manner-render-sh > renders/$(basename ${f}).html
-done
+
+if [ "$#" -eq 1 ]; then
+  renderFile $1
+else
+  rm renders/*
+  for f in ../manlib/man/man1/*; do
+    renderFile ${f}
+  done
+fi
